Fixed migrations

+++ database/migrations/2020_01_31_102828_create_colors_table.php
@@ -15,7 +15,7 @@ class CreateColorsTable extends Migration
     {
         Schema::create('colors', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('palette_id');
+            $table->bigInteger('palette_id');
             $table->tinyInteger('sort');
             $table->string('color', 6);
